refactor: reusable shared card cover
